chore: fix types + utils file-format & code-split
