+++ apps/dashboard/src/components/release-pipeline/types/pipeline-sidebar.props.ts
